requirements.txt
flask==3.0.2
anyio==4.8.0
certifi==2025.1.31
exceptiongroup==1.2.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.10
packaging==24.2
pillow==11.1.0
pytesseract==0.3.13
python-dotenv==0.21.0
rfc3986==1.5.0
sniffio==1.3.1
typing_extensions==4.12.2
quart==0.19.4
uvicorn==0.34.0
python-telegram-bot[webhooks]==20.0

